Delete Authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
 
@@ -24,7 +26,7 @@ class ListingController {
   public function index() 
   {
 
-    $listings = $this->db->queryWithShareLock('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db->queryWithShareLock('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
   // inspect($listings); tes
 
@@ -83,7 +85,7 @@ class ListingController {
 
     $newListingData = array_intersect_key($_POST,array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     $newListingData = array_map('sanitize', $newListingData);
 
@@ -153,10 +155,19 @@ class ListingController {
 
     $listing = $this->db->queryWithShareLock('SELECT * FROM listings WHERE id = :id', $params)->fetch();
     
+    // Check if listing exists
     if(!$listing) {
       ErrorController::notFound('Listing not found');
       return;
-    } 
+    }
+
+    // Authorization
+    if(!Authorization::isOwner($listing->user_id)) {
+      $_SESSION['error_message'] = 'You\'re not authorized to delete this listing';
+      return redirect('/listings/' . $listing->id);
+
+    }
+
       $this->db->queryWithShareLock('DELETE FROM listings WHERE id = :id', $params);
 
       // Set flash message
